refactoring

+++ rating/src/main/java/org/example/rating/dto/read/RideReadDto.java
@@ -1,15 +1,13 @@
 package org.example.rating.dto.read;
 
-import java.math.BigDecimal;
-
 public record RideReadDto(
-        Long id,
-        Long driverId,
-        Long passengerId,
+        long id,
+        long driverId,
+        long passengerId,
         String addressFrom,
         String addressTo,
         String driverRideStatus,
         String passengerRideStatus,
-        BigDecimal cost
+        double cost
 ) {
 }
